Bash script doesn't download files to folders but downloads a bunch of urls as filenames.
I have created a script for Bash that will save artifacts using the path/url in the array element. This saves the artifacts to the current folder and keeps dirs intact. At least this is the behavior on the terminal in OSX 10.14.6. Feel free to refactor this to a single line or smaller code.